pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speechbubble"
version = "0.1.0"
description = "Automatic speech bubble insertion for illustration workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.10",
  "Pillow>=10.0",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "pydantic>=2.8",
  "gradio>=4.44",
  "huggingface-hub>=0.33.5,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "httpx>=0.27",
  "ruff>=0.6",
]

[project.scripts]
speechbubble = "speechbubble.cli:main"
speechbubble-api = "speechbubble.api:run"
speechbubble-ui = "speechbubble.ui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
